+++ woninet/exc.py
@@ -1,0 +1,2 @@
+class PingUtilityNotFound(Exception):
+    """Raised if the ping command is not found on PATH."""