+++ CHANGELOG.md
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+### Added
+- Added custom `PingUtilityNotFound` exception for when the `ping` command is not found on the `PATH`.
+- Added the `PingResult` dataclass to preserve `system_ping()`'s output
+- Added `system_ping()` as fallback for `icmplib.ping()`, when `icmplib.ping()` raises `SocketPermissionError`, which happends for root privileges issues.
+- Added `ping_method` variable to cache the ping method, to avoid repeated fallback overhead.
+- Added test for parsing `system_ping()`'s output for Linux and Windows operating systems.
+- Added warning message when Privileged ICMP is unavailable, at startup.
+
 ### Changed
 
 - Disable ruff formatting in `create_configuration.py` module to preserve the `JSON` structure.